refactor(AnswerRepositoryTest, AnswerTest): 통합 테스트(DB 연결)와 도메인 단위 테스트 분리
